ARC32 Select 4Kb MMU pages in uClibc
uClibc assumes default MMU size for ARC processors of 8Kb
which was a reasonable choice for ARCompact (AKA ARC ISAv1) &
ARCv2 ISA processors.

In case of ARCv3 ISA processors default value is 4 Kb, moreover
32-bit flavor of ARCv3 processors ARC HS58 may only have 8 Kb pages.

And so to make sure proper page size is used, we need to propagate
that non-default setting to uClibc. Which we do here.

Note, there's no check for 64-bit flavor of ARCv3 processors
in that change, that's because currently there're no plans for
adding support for them into uClibc, instead glibc must be used.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>